pyproject.toml
[project]
name = "d2r-gear-optimiser"
version = "0.1.0"
description = "D2R gear loadout optimiser — exhaustive search across inventory, runes, jewels, runewords."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlmodel>=0.0.22",
    "click>=8.1",
    "rich>=13.9",
    "pyyaml>=6.0",
    "pydantic>=2.10",
    "google-genai>=1.22.0",
    "openai>=1.54.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "alembic>=1.14",
]

[project.scripts]
optimise = "d2r_optimiser.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/d2r_optimiser"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "alembic>=1.18.4",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
]
